Splits SocksServer classes into it's own files
