+++ proxy/SocksServer.h
@@ -1,0 +1,47 @@
+/** @file
+
+  A brief file description
+
+  @section license License
+
+  Licensed to the Apache Software Foundation (ASF) under one
+  or more contributor license agreements.  See the NOTICE file
+  distributed with this work for additional information
+  regarding copyright ownership.  The ASF licenses this file
+  to you under the Apache License, Version 2.0 (the
+  "License"); you may not use this file except in compliance
+  with the License.  You may obtain a copy of the License at
+
+      http://www.apache.org/licenses/LICENSE-2.0
+
+  Unless required by applicable law or agreed to in writing, software
+  distributed under the License is distributed on an "AS IS" BASIS,
+  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  See the License for the specific language governing permissions and
+  limitations under the License.
+ */
+
+#pragma once
+
+class ParentConfigParams;
+
+/*
+  For supporting multiple Socks servers, we essentially use the
+  ParentSelection infrastructure. Only the initialization is different.
+  If needed, we will have to implement most of the functions in
+  ParentSection.cc for Socks as well. For right now we will just use
+  ParentSelection
+
+  All the members in ParentConfig are static. Right now
+  we will duplicate the code for these static functions.
+*/
+struct SocksServerConfig {
+  static void startup();
+  static void reconfigure();
+  static void print();
+
+  static ParentConfigParams *acquire();
+  static void release(ParentConfigParams *params);
+
+  static int m_id;
+};
